fixed #977, eliminated compiler warnings (all, but one)
git-svn-id: http://sphinxsearch.googlecode.com/svn/branches/rel20/@3008 8b96e2b9-35c5-2c16-bc47-5122d61876d4

+++ src/sphinxint.h
@@ -1366,7 +1366,7 @@ public:
 		// how many bytes of a previous keyword can we reuse?
 		BYTE iMatch = 0;
 		int iMinLen = Min ( m_iLastLen, iLen );
-		assert ( iMinLen<sizeof(m_sLastKeyword) );
+		assert ( iMinLen<(int)sizeof(m_sLastKeyword) );
 		while ( iMatch<iMinLen && m_sLastKeyword[iMatch]==pWord[iMatch] )
 		{
 			iMatch++;
